contivctl,netctl: rename contivctl to netctl, for github.com/contiv/contivctl
